Renames lib/google-apps to lib/gsuite
