+++ client/me/purchases/product-link/index.jsx
@@ -15,7 +15,7 @@ import i18n from 'i18n-calypso';
 import config from 'config';
 import { domainManagementEdit } from 'my-sites/domains/paths';
 import { getThemeDetailsUrl } from 'state/themes/selectors';
-import { googleAppsSettingsUrl } from 'lib/google-apps';
+import { googleAppsSettingsUrl } from 'lib/gsuite';
 import {
 	isDomainProduct,
 	isGoogleApps,
